build-sys: Use git-version-gen to get the version from git
This will keep generating version numbers identical to the ones we are
currently using if built from a clean tree on a tagged commit.

If built from a tree with additional commits (or from a dirty tree),
this will produce a version number that reflects the information from
`git describe`, e.g. 225.221-b47c-dirty if building on b47c788854600c
with uncommitted modifications on top.

The git-version-gen script is developed in the gnulib repository:
http://git.savannah.gnu.org/gitweb/?p=gnulib.git;a=blob;f=build-aux/git-version-gen;h=8e92c0ad8e3db950ef1cc1a3a19fee68cdddb77f;hb=HEAD

Distribute it from tools/ (instead of build-aux/) since we have a
.gitignore rule for the whole of build-aux/.

Followed the instructions on the script for the necessary modifications
to configure.ac and Makefile.am (I skipped the part about the .version
file, since that breaks `distcheck` and we do not seem to require it.)

Tested:
- Built from a dirty tree on top of an untagged commit.
- Tagged v226 on top of this commit, built on it, confirmed the
  generated version was indeed "systemd 226".
- Built tarballs with `make dist` and confirmed that building from
  the unpacked contents of the tarball works and produces the expected
  version (same version as a build on the git tree would produce.)
- Booted the system and confirmed that the initialization message on the
  journal includes the full version ("systemd 225.221-b47c-dirty running
  in system mode ..."), also `systemctl --version` will produce the
  complete version number.
- No regressions in tests, in particular in `make distcheck`.