delta: support instantiated units and dropins
Consider the following filesystem contents:

  /usr/lib/systemd/system/httpd.service
    env.conf
  /usr/lib/systemd/system/postfix.service
  /usr/lib/systemd/system/getty@.service
  /usr/lib/systemd/system/getty@.service.d/
    ext.conf
  /etc/systemd/system/getty@tty1.service.d/
    test1-a.conf
    test1-b.conf
  /etc/systemd/system/getty@tty2.service
    test2-c.conf
    test2-d.conf
  /etc/systemd/system/getty@tty2.service.d/
  /run/systemd/system/getty@tty5.service

The resulting internal representiation will be as follows.

   top[httpd.service] = /usr/lib/systemd/system/httpd.service
   bottom[httpd.service]
     [/usr/lib/systemd/system/httpd.service] -> env.conf

   top[postfix.service] = /usr/lib/systemd/system/postfix.service
   bottom[postfix.service]
     [/usr/lib/systemd/system/postfix.service] -> (empty list)

   top[getty@.service] = /usr/lib/systemd/system/getty@.service
   bottom[getty@.service]
     [/usr/lib/systemd/system/getty@.service]  -> ext.conf
     [/etc/systemd/system/getty@tty1.service]  -> test1-a.conf, test1-b.conf
     [/etc/systemd/system/getty@tty2.service]  -> test2-c.conf, test2-d.conf
     [/run/systemd/systemd/getty@tty5.service] -> (empty list)

The 'top' hashmap contains the filename of the very first unit, which is
considered as the basis for extension. It's key is a base name of the
full path of the unit it holds as its value, or basename of an template
unit.

The 'bottom' hashmap holds a hashmap with override-to-extension
mappings and uses the same key as the 'top' hashmap. The extensions are
saved in a plain string list. Empty list means that there are only
overrides to the corresponding unit in 'top'.