Add utf8->utf16/ucs2 functions conversion
Add 4 functions to convert from a utf8 encoded string to utf16 or ucs2 (
utf16 limited to the so called BMP plabe) encoded string. The functions
return the new allocated string and its length (in bytes). The input string
is assumed zero terminated; for the functions with the _len
suffix, it is possible to pass the length of the source utf8 string
to check if this is longer than the allowed.

If the utf8 string is bad encoded, or it represents not unicode char, or it
is more longer than the limit, NULL is returned.

Two more tests are added to test-utf8.c.