bootctl: set-boot-entry command, set the default entry for the bootloader
Add "set-boot-entry" command to set the boot entry. The syntax is:

	# bootctl set-boot-entry [one-shot] <bootentry>

If "one-shot" is passed, the entry is used as boot entry only for next boot.
<bootentry> may be:
  - the path of the config file, like:
    /boot/loader/entry/e84907d099904117b355a99c98378dca-4.2.0.conf.
  - the entry number as listed by the command "list-entries"
  - "default", which selected the 'most recent' entry
  - the name of the entry, which is the file name without the .conf (eg:
    the path above would be e84907d099904117b355a99c98378dca-4.2.0)


NOTE: passing the *absolute path* is the only un-ambigous way of referring an
entry. The other methods use some heuristichs to select and entry: if
the entry filename is "default.conf" or "1.conf", unexpected results may
happen.
NOTE2: UEFI deals only with ucs2 characters, so pay attention that the file
name/entry contains onlu ucs2 valid character (unicode BMP plane)
NOTE3: "default" and "entry number" are based by the sort of the entries. The
sorting alghoritm is the same of the list-entries command.