Ask for passphrases not only on the first console of /dev/console
but also on all other consoles.  This does help on e.g. mainframes
where often a serial console together with other consoles are
used.  Even rack based servers attachted to both a serial console
as well as having a virtual console do sometimes miss a connected
monitor.

For each console device fork a inquiring task for passphrase and do
not return to the caller.  To be able to ask on all terminal devices
of /dev/console the devices are collected. If more than one device
are found, then on each of the terminals a inquiring task is
forked.

Every task has its own session and its own controlling terminal.
If one of the tasks does handle a password, the remaining tasks
will be terminated.

Also let contradictory options on the command of
systemd-tty-ask-password-agent fail.