sd-id128: add sd_id128_get_machine() call to rerieve the machine secret
The machine secret is modelled after machine-id and is used for purposes where
the actual machine secret needs to remain non-public.

In particular, this will be used by network management tooling to seed the
RFC7217 address generator that is designed to generate stable addresses that
are not trackable when the machine moves across different networks.