bootctl: Handle gummiboot -> systemd migration
On Linux systems we may encounter cases where the legacy gummiboot binary
is still present on the ESP. In these cases a `bootctl install` or even
a `bootctl upgrade` will fail on files such as BOOTX64.EFI, as their internal
version string will not match that of systemd.

To counteract this upgrade problem, we will always continue to replace the
legacy gummiboot (<=48) with systemd's merged tree, to provide a valid upgrade
route for distributions.