Generalize CK_COND sense data checking
Check sense key and additional sense code as per SAT instead of ATA status return descriptor header.

Some devices return "fixed format sense data" which are not strictly compliant with SAT, but that doesn't stop the ATA identify/passthrough from being complete and error-free, so we have check for the data as well which will allow the device to export their info. 

Actually as of Revision 02 (sat4r02.pdf), draft SAT-4 have a section "12.2.2.7 Fixed format sense data", and according to the footnote "a" in Table 127, "SATLs compliant with versions prior to SAT-2 return descriptor format sense data for the ATA PASS-THROUGH commands regardless the D_SENSE bit", which seem to hint that SATLs compliant with later versions of SAT is allowed to return fixed format sense data.